Restructuring and renaming.
