+++ lua/diffview/api/c_view.lua
@@ -1,9 +1,9 @@
 local oop = require'diffview.oop'
 local utils = require'diffview.utils'
 local EventEmitter = require'diffview.events'.EventEmitter
-local View = require'diffview.view'.View
-local CFileEntry = require'diffview.api.c-file-entry'.CFileEntry
-local FilePanel = require'diffview.file-panel'.FilePanel
+local View = require'diffview.scene.view'.View
+local CFileEntry = require'diffview.api.c_file_entry'.CFileEntry
+local FilePanel = require'diffview.scene.file_panel'.FilePanel
 local FileDict = require'diffview.git'.FileDict
 local Rev = require'diffview.rev'.Rev
 local RevType = require'diffview.rev'.RevType
